+++ packages/workato-adapter/jest.config.js
@@ -26,8 +26,8 @@ module.exports = deepMerge(
     ],
     coverageThreshold: {
       'global': {
-        branches: 83.5,
-        functions: 94,
+        branches: 85,
+        functions: 95,
         lines: 95,
         statements: 95,
       },
