+++ packages/workato-adapter/src/reference_mapping.ts
@@ -106,13 +106,6 @@ export const fieldNameToTypeMappingDefs: FieldReferenceDefinition[] = [
     serializationStrategy: 'id',
     target: { type: 'connection' },
   },
-
-  // serialization-only rule, for translating the reference back to the original value pre-deploy
-  // (not used in fetch-only)
-  {
-    src: { field: 'code', parentTypes: ['recipe'] },
-    serializationStrategy: 'fullValue',
-  },
 ]
 
 // We can extract the api name from the elem id as long as we don't support renaming
